requirements.txt
yfinance
apache-airflow-providers-postgres
requests
vaderSentiment
psycopg2-binary
pandas
streamlit
matplotlib
